requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.12.0

# API clients
pinecone-client>=2.2.0
openai>=1.0.0

# Visualization libraries
plotly>=5.11.0
mplfinance>=0.12.0
kaleido>=0.2.1  # For plotly static image export

# Technical analysis
ta>=0.10.0

# Environment management
python-dotenv>=0.19.0

# Jupyter support (optional)
notebook>=6.4.0
ipywidgets>=7.6.0

# Additional utilities
tqdm>=4.64.0  # Progress bars
requests>=2.28.0